+++ RELEASENOTES.md
@@ -2,7 +2,7 @@
 
 ### Version 1.1.1
 Release date 2021-04-22
-- **External Freight Customers** overview including supplied and consumed cargo types.
+- **Freight Customers** overview including supplied and consumed cargo types.
 - **Bug fixes** in external *freight customers*
 -
 ### Version 1.1.0