+++ RELEASENOTES.md
@@ -4,7 +4,7 @@
 Release date 2021-04-22
 - **Freight Customers** overview including supplied and consumed cargo types.
 - **Bug fixes** in external *freight customers*
--
+
 ### Version 1.1.0
 Release date 2021-04-21
 - **External stations** to describe real world stations with *freight customers* and *cargo flow*.